import functions from helper.js
