+++ index.html
@@ -4,6 +4,7 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <link rel="stylesheet" href="./style.css" />
+    <script type="module" src="./index.js"></script>
     <title>Todo List</title>
   </head>
   <body>
@@ -13,6 +14,5 @@
       <div class="todos">
       </div>
     </div>
-    <script src="./index.js"></script>
   </body>
 </html>
